Arranged extensions in alphabetical order.

+++ Sources/BFKit/Linux/Foundation/NumberExtension.swift
@@ -99,9 +99,10 @@ public func randomFloat() -> Float {
 public func randomFloat(min minValue: Float, max maxValue: Float) -> Float {
     return randomFloat() * abs(minValue - maxValue) + min(minValue, maxValue)
 }
+
 // MARK: - Extensions
 
-public extension Int {
+public extension Double {
     /// Gets the individual numbers, and puts them into an array. NOTE: All negative numbers will start with 0.
     var array: [Int] {
         return description.characters.map{Int(String($0)) ?? 0}
@@ -115,7 +116,7 @@ public extension Float {
     }
 }
 
-public extension Double {
+public extension Int {
     /// Gets the individual numbers, and puts them into an array. NOTE: All negative numbers will start with 0.
     var array: [Int] {
         return description.characters.map{Int(String($0)) ?? 0}
